+++ packages/admin-ui/ui/src/i18n/index.ts
@@ -10,8 +10,20 @@ export const supportedLanguages = [
     name: "English",
   },
   {
+    locale: "de",
+    name: "Deutsch",
+  },
+  {
     locale: "fr",
     name: "Français",
+  },
+  {
+    locale: "hr",
+    name: "Hrvatski",
+  },
+  {
+    locale: "it",
+    name: "Italiano",
   },
   {
     locale: "pt",
@@ -29,6 +41,14 @@ export const supportedLanguages = [
     locale: "bn",
     name: "বাংলা (Bangladesh)",
   },
+  {
+    locale: "es",
+    name: "Español",
+  },
+  {
+    locale: "bs",
+    name: "Bosanski",
+  }
 ]
 
 const adminPath = getFullAdminPath()